LayoutTests/media/progress-event-total-expected.txt

EVENT(loadstart)
EXPECTED (!event.lengthComputable == 'true') OK
EXPECTED (event.total == '0') OK
EXPECTED (event.loaded == '0') OK
EVENT(load)
EXPECTED (event.lengthComputable == 'true') OK
EXPECTED (event.loaded > '0') OK
EXPECTED (event.total == '239528') OK
END OF TEST
